Split static components of ActionControl to api
