+++ GEECS-Scanner-GUI/geecs_scanner/app/lib/action_control.py
@@ -3,7 +3,7 @@ Contains a simpler version of a full "RunControl" that is only capable of perfor
 
 -Chris
 """
-from typing import Union, Any
+from typing import Any
 from geecs_scanner.data_acquisition import ActionManager
 import logging
 
@@ -21,68 +21,6 @@ class ActionControl:
         :param experiment_name: Name of the experiment, used to find the right config files and connect to devices
         """
         self.action_manager = ActionManager(experiment_dir=experiment_name)
-
-    @staticmethod
-    def get_new_action(action) -> Union[None, dict[str, str]]:
-        """
-        Translates a given action keyword to a default dictionary for that action keyword
-
-        :param action: action keyword
-        :return: default dictionary for the associated action
-        """
-        default = None
-        if action == 'set':
-            default = {
-                'action': 'set',
-                'device': '',
-                'variable': '',
-                'value': ''
-            }
-        elif action == 'get':
-            default = {
-                'action': 'get',
-                'device': '',
-                'variable': '',
-                'expected_value': ''
-            }
-        elif action == 'wait':
-            default = {
-                'wait': ''
-            }
-        elif action == 'execute':
-            default = {
-                'action': 'execute',
-                'action_name': ''
-            }
-        elif action == 'run':
-            default = {
-                'action': 'run',
-                'file_name': '',
-                'class_name': ''
-            }
-        return default
-
-    # List of available actions, to be used by the completer for the add action line edit
-    list_of_actions = [
-        'set',
-        'get',
-        'wait',
-        'execute',
-    ]
-
-    @staticmethod
-    def generate_action_description(action: dict[str, list]) -> str:
-        """ For each action type, generate a string that displays all the information for that action step """
-        description = "???"
-        if action.get("wait") is not None:
-            description = f"wait {action['wait']}"
-        elif action['action'] == 'execute':
-            description = f"execute {action['action_name']}"
-        elif action['action'] == 'set':
-            description = f"{action['action']} {action['device']}:{action['variable']} {action.get('value')}"
-        elif action['action'] == 'get':
-            description = f"{action['action']} {action['device']}:{action['variable']} {action.get('expected_value')}"
-        return description
 
     def perform_action(self, action_list: dict[str, list]):
         """
